Initialize P2P on p2p0 and not wlan0 (JB)
P2P is being initialized using a global struct
which points to the first initialized interface wpa_s
struct. That's causing P2P commands sent on the control
socket matching p2p0 to generate scan and ROC on the
wrong interface (wlan0) as well as events like PEER_DEVICE_FOUND
are sent on the wrong control socket to the Android framework.
This caused the framework to ignore these and prevent
displaying the P2P peer in the UI.

Solve it by an ugly hack for now to enable P2P only when the
interface p2p0 is being initialized.

This probably should be fixed upstream as well by marking
somehow on which interface you can do p2p or enable p2p per
interface.

[Basic patch needed for p2p to work in JB UI]

Change-Id: I191b94ba5983c6f83c8653655029c8def6b43d74
Signed-off-by: Eyal Shapira <eyal@wizery.com>
Signed-off-by: Vishal Mahaveer <vishalm@ti.com>